Moved add_donator and rm_donator to admin commands; template for weather set
