+++ helpers/commandsList.py
@@ -20,13 +20,14 @@ commands_list = {
     "osuset": commands.osuCommands.OsuSet,
     "role": commands.donatorCommands.GetRole,
     "роль": commands.donatorCommands.GetRole,
-    "add_donator": commands.donatorCommands.AddDonator,
-    "rm_donator": commands.donatorCommands.RemoveDonator,
+    "add_donator": commands.adminCommands.AddDonator,
+    "rm_donator": commands.adminCommands.RemoveDonator,
     "op": commands.adminCommands.Op,
     "deop": commands.adminCommands.Deop,
     "mm": commands.osuCommands.MatchmakingStats,
     "top": commands.osuCommands.TopScoreCommand,
     "last": commands.osuCommands.RecentScoreCommand,
+    "ласт": commands.osuCommands.RecentScoreCommand,
     "compare": commands.osuCommands.Compare,
     "c": commands.osuCommands.Compare,
     "с": commands.osuCommands.Compare,
